Refactored ListenPoolState to a separate module.
