+++ components/channeler/src/lib.rs
@@ -8,6 +8,7 @@
 mod channeler;
 mod types;
 mod transform_pool;
+mod listen_pool_state;
 mod listen_pool;
 mod connect_pool;
 mod connector_utils;
